moved redux hub to src level (out of hooks)
