+++ index.js
@@ -7,7 +7,7 @@ import {AppRegistry} from 'react-native';
 import App from './App';
 import {name as appName} from './app.json';
 import {Provider} from 'react-redux';
-import store from './src/Hooks/ReducerHub/store';
+import store from './src/ReducerHub/store';
 const AppRedux = () => (
   <Provider store={store}>
     <App />
