+++ src/ReducerHub/store.tsx
@@ -7,13 +7,9 @@ import SagaData from './saga';
 
 const sagaMiddleware = createSagaMiddleware();
 
-
-
 const store = configureStore({
   reducer: rootReducer,
-  // middleware: getDefaultMiddleware =>
-  //   getDefaultMiddleware().concat(sagaMiddleware),
-  middleware: (getDefaultMiddleware) =>
+  middleware: getDefaultMiddleware =>
     getDefaultMiddleware({
       thunk: false,
       immutableCheck: true,
